# #25251: improve tile exchange in ttnn sort