### Ticket

[#25251](https://github.com/tenstorrent/tt-metal/issues/25251)

### Problem description
- Tests have shown that the biggest bottleneck in the current implementation of the Cross Core Data Exchange strategy in ttnn sort is communication between kernels within a single core. In order to reduce this, two-tile transfer was added so that every second tile did not have to wait for transfer between cores, but was exchanged when the first pair was sorted.

### What's changed
- Added two-tile transfer between kernels.
- Changed copy tile between two CB functionality.

### Performance

Measured on Wormhole

Tiles | Cores used | Tiles per core | Tile Width | DEVICE KERNEL DURATION PER CORE MIN [ms] | DEVICE KERNEL DURATION PER CORE MAX [ms] | DEVICE KERNEL DURATION PER CORE AVG [ms] | Bandwidth MElem/s
-- | -- | -- | -- | -- | -- | -- | --
Current implementation
64 | 32 | 2 | 32 | 0.176 | 0.408 | 0.402 | 11.609
128 | 64 | 2 | 32 | 0.241 | 0.178 | 0.178 | 16.987
256 | 64 | 4 | 32 | 0.553 | 0.563 | 0.558 | 14.821
512 | 64 | 8 | 32 | 1.223 | 1.244 | 1.236 | 13.392
1024 | 64 | 16 | 32 | 2.746 | 2.788 | 2.769 | 11.934
2048 | 64 | 32 | 32 | 6.187 | 29.044 | 28.819 | 10.592
4096 | 64 | 64 | 32 | 13.266 | 6.271 | 6.232 | 9.880
8192 | 64 | 128 | 32 | 28.242 | 28.578 | 28.426 | 9.282
Transfer two tiles at a time
64 | 32 | 2 | 32 | 0.157 | 0.159 | 0.158 | 13.040
128 | 64 | 2 | 32 | 0.216 | 0.220 | 0.218 | 18.978
256 | 64 | 4 | 32 | 0.453 | 0.462 | 0.457 | 18.095
512 | 64 | 8 | 32 | 1.090 | 1.110 | 1.101 | 15.034
1024 | 64 | 16 | 32 | 2.429 | 2.469 | 2.451 | 13.488
2048 | 64 | 32 | 32 | 5.248 | 5.405 | 5.334 | 12.489
4096 | 64 | 64 | 32 | 12.568 | 12.874 | 12.736 | 10.429
8192 | 64 | 128 | 32 | 28.121 | 29.447 | 29.167 | 9.322


### Checklist
- [ ] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes
- [ ] [Blackhole Post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml) CI with demo tests passes (if applicable)
- [ ] [Model regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable)
- [ ] [Device performance regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml) CI passes (if applicable)
- [ ] (For models and ops writers) [Single-card demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml) CI passes (if applicable) See [recommended dev flow](https://github.com/tenstorrent/tt-metal/blob/main/models/MODEL_ADD.md#a-recommended-dev-flow-on-github-for-adding-new-models).
- [ ] [Galaxy quick](https://github.com/tenstorrent/tt-metal/actions/workflows/tg-quick-trigger.yaml) CI passes (if applicable)
- [ ] [TG demo tests, for Llama](https://github.com/tenstorrent/tt-metal/actions/workflows/tg-demo-tests.yaml) CI passes, if applicable, because of current Llama work
- [ ] (For runtime and ops writers) [T3000 unit tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-unit-tests.yaml) CI passes (if applicable, since this is run on push to main)
- [ ] (For models and ops writers) [T3000 demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-demo-tests.yaml) CI passes (if applicable, since this is required for release)
- [ ] (For models and ops writers) [Single-card demo tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml) CI passes (if applicable, since this is required for release)
- [ ] New/Existing tests provide coverage for changes